requirements.txt
lxml
pycryptodomex
PyExecJS
requests